Remove debugs, we seem to have found it.
The first issue was that Sphinx is returning array(float) instead of
just a float for energies.

The second issue was that the numpy comparer didn't have the relative
tolerance set, so even though the absolute tolerance was very strict,
two clearly different values were being evaluated as the same because
their relative difference was fairly small.