Make GenericJob-inheriting protocols special classes
With Dominik's dynamic typing, I had issues with pyiron not being able
to reload the jobs from HDF. I don't see an easy way around this, so
here the main meat of each algorithm is abstracted and only inherits
from CompoundVertex, and the Protocol class is now reserved for the top-
level graph being instantiated as a pyiron job. Each algorithm, 'X' has
its basic logic, and an extra 'ProtocolX' class which can be
instantiated as a job.

I had to play around with the init methods of the Vertices to avoid
using super since there are now signature mismatches. Other than that,
I'm at least satisfied with this solution; it's not as snazzy as
dynamic typing, but it's not clear to me how to get that to mesh with
pyiron's storage and retrieval...