Black magic
Multiple inheritance in python is a mess. I changed my mind about
throwing away the use of super when I realized that switching to
explicit parent calls would also be needed in other children like
ListVertex. There was a bit of wrangling necessary to get the super
calls working though, since the signatures of the Vertex-based code and
GenericJob-based code are different. I also had to stop overloading the
use of the name attribute, so vertices now use vertex_name as an
attribute. Integration tests including saving and loading all pass...