Refactor hdf code
It can be shorter. With group_name it was wrapped in with, and
otherwise not. Any necessary close call must be taken care of later, and
adding an explicit close here created some sort of database orphan,
hence no close calls despite the fact we got rid of a with.