Fix problems caused by clock
When the job is submitted to the queue, somehow, `clock.output.n_counts[-1]' always remains at its initialized value of 0 when accessed during the run. It goes to the correct value at the end. For now, I pass `total_counts' (I use this in `running_average_positions' within the CompoundVertex to keep track of the parent protocol count, in order to account for thermalization without having to have a seperate node for it in the CompoundVertex) to `check_thermalized', which always returned False, and thus bypassed all th string processing nodes. I will discuss this issue with Liam, if I do not resolve it myself first.