Rebase to inherit from FlexibleMaster and ParallelMaster
It doesn't consistently work like it looks like it should but there are workarounds. Also the HDF management is still not ideal. I particularly did not find a good way to handle the fact that there are three rather than one reference job, especially when it all gets wrapped in another layer of listing.