Force output collection
In the long run we should reconsider the setup, but for now we can just force this FlexibleMaster child to collect it's output. This is in preparation for getting the current output class to write to HDF. I'm a bit sad about how that duplicates data and burns disk space, but it synchronizes it with the other job classes and allows inpsection. Ideal would be some sort of symbolic linking to the other parts of the HDF, but I don't know how to do that atm and will leave such an improvement for later.