Re-add HarmonicMD
The HessianJob calculates the restoring forces for static MD positions, and needs to run in a loop over MD snapshots of the atom positions. This protocol is independent of the MD run that the HessianJob needs, and uses Langevin Dynamics independent of a Calculator (Lammps, VASP etc.). I will leave this here, as it this protocol presently is faster than running HessianJob over a loop.