Move the protocol notebooks to a new directory
The notebook tests were failing because lammps wasn't one of the dependencies. Considering that the notebooks use lammps, and in one case use vasp as well, I move them to a new directory, so the tests don't run. This however makes it even more pressing to have unit tests written for the protocols and their associated vertices.