assign core to child execution
Turns out that the error that I saw in the free_energy_class branch, while executing the FreeEnergy job persisted even in this branch. In summary, as I increased the number of lambdas, the execution just hanged (hung?). I thought that multiprocessing.Process automatically assigned a core to a lambda process, but after our discussion with Joerg, I was not sure anymore. On further digging, it seemed like it was just sharing the cores, and not using a core per lambda process, like I had initially thought.  As Joerg suggested, I made sure to isolate the execution of a lambda process to a specific core, following this nice little reference: https://stackoverflow.com/questions/5784389/using-100-of-all-cores-with-the-multiprocessing-module/35371568#35371568

It worked on my local machine: I have 6 cores and 12 threads, and I could for sure see that 2 cores were FAIRLY idle. Time to test it out on the cluster.