Add creators for UX
Add creator classes for jobs, tasks, structures and executors so that
users don't need to import them.  With it there are some changes in
TinyJob:

1. it is no longer an abstract class, but requires a task instance to be
   instantiated.  Reloading already run jobs can be done with
   TinyJob.restore
2. as a consequence GenericTinyJob is no longer needed
3. because Executors are easily created now, the _executors class
   attribute is also gone; TinyJob.run now takes an argument 'executor',
   either an executor or a method name of the ExecutorCreator as a short
   hand

The creators are very similar to the ones already existing in
pyiron_base.  Notable is the ExecutorCreator which as a method
`most_recent` that simply returns the last created executor, which is
also used by TinyJob.run by default, freeing the user from lugging
around executor instances on their own.