Add generic job factories
A JobFactory is a light weight object that represents a certain type of
pyiron job.  It can be used by abstract workflows to run jobs without
knowing what the reference model and the detailed parameters e.g. for
the HPC queue are, without having to pass all of this information
through all layers of the workflow.